+++ src/main/java/org/devkor/apu/saerok_server/domain/auth/application/AbstractSocialAuthService.java
@@ -9,29 +9,24 @@ import org.devkor.apu.saerok_server.domain.auth.application.facade.AuthTokenFaca
 import org.devkor.apu.saerok_server.domain.auth.core.entity.SocialAuth;
 import org.devkor.apu.saerok_server.domain.auth.core.entity.SocialProviderType;
 import org.devkor.apu.saerok_server.domain.auth.core.repository.SocialAuthRepository;
-import org.devkor.apu.saerok_server.domain.auth.support.SocialAuthClient;
-import org.devkor.apu.saerok_server.domain.auth.support.SocialUserInfo;
-import org.devkor.apu.saerok_server.domain.user.core.entity.User;
-import org.devkor.apu.saerok_server.domain.user.core.entity.UserRole;
-import org.devkor.apu.saerok_server.domain.user.core.entity.UserRoleType;
-import org.devkor.apu.saerok_server.domain.user.core.repository.UserRepository;
-import org.devkor.apu.saerok_server.domain.user.core.repository.UserRoleRepository;
+import org.devkor.apu.saerok_server.domain.auth.core.service.UserProvisioningService;
+import org.devkor.apu.saerok_server.domain.auth.infra.SocialAuthClient;
+import org.devkor.apu.saerok_server.domain.auth.core.dto.SocialUserInfo;
 import org.devkor.apu.saerok_server.global.util.dto.ClientInfo;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 
 @Slf4j
 @RequiredArgsConstructor
+@Transactional
 public abstract class AbstractSocialAuthService {
 
     private final SocialAuthRepository socialAuthRepository;
-    private final UserRepository userRepository;
-    private final UserRoleRepository userRoleRepository;
     private final AuthTokenFacade authTokenFacade;
+    private final UserProvisioningService userProvisioningService;
 
     protected abstract SocialAuthClient client();
 
-    @Transactional
     public ResponseEntity<AccessTokenResponse> authenticate(String authorizationCode, String accessToken, ClientInfo ci) {
 
         SocialUserInfo userInfo = client().fetch(authorizationCode, accessToken);
@@ -40,7 +35,7 @@ public abstract class AbstractSocialAuthService {
 
         SocialAuth socialAuth = socialAuthRepository
                 .findByProviderAndProviderUserId(provider, userInfo.sub())
-                .orElseGet(() -> provisionNewUser(provider, userInfo));
+                .orElseGet(() -> userProvisioningService.provisionNewUser(provider, userInfo));
 
         AuthBundle bundle = authTokenFacade.issueTokens(
                 socialAuth.getUser(),
@@ -51,14 +46,5 @@ public abstract class AbstractSocialAuthService {
                 .ok()
                 .header(HttpHeaders.SET_COOKIE, bundle.cookie().toString())
                 .body(bundle.body());
-    }
-
-    private SocialAuth provisionNewUser(SocialProviderType provider, SocialUserInfo userInfo) {
-        User user = userRepository.save(User.createUser(userInfo.email()));
-        userRoleRepository.save(UserRole.createUserRole(user, UserRoleType.USER));
-
-        return socialAuthRepository.save(
-                SocialAuth.createSocialAuth(user, provider, userInfo.sub())
-        );
     }
 }
