+++ src/main/java/org/devkor/apu/saerok_server/domain/auth/core/dto/SocialUserInfo.java
@@ -1,4 +1,4 @@
-package org.devkor.apu.saerok_server.domain.auth.support;
+package org.devkor.apu.saerok_server.domain.auth.core.dto;
 
 public record SocialUserInfo(
         String sub,
