+++ src/main/java/org/devkor/apu/saerok_server/domain/auth/infra/KakaoAuthClient.java
@@ -5,8 +5,7 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 import lombok.RequiredArgsConstructor;
 import org.devkor.apu.saerok_server.domain.auth.core.entity.SocialProviderType;
 import org.devkor.apu.saerok_server.domain.auth.infra.dto.KakaoUserInfoResponse;
-import org.devkor.apu.saerok_server.domain.auth.support.SocialAuthClient;
-import org.devkor.apu.saerok_server.domain.auth.support.SocialUserInfo;
+import org.devkor.apu.saerok_server.domain.auth.core.dto.SocialUserInfo;
 import org.devkor.apu.saerok_server.global.exception.UnauthorizedException;
 import org.springframework.stereotype.Service;
