+++ src/main/java/org/devkor/apu/saerok_server/domain/auth/infra/SocialAuthClient.java
@@ -1,5 +1,6 @@
-package org.devkor.apu.saerok_server.domain.auth.support;
+package org.devkor.apu.saerok_server.domain.auth.infra;
 
+import org.devkor.apu.saerok_server.domain.auth.core.dto.SocialUserInfo;
 import org.devkor.apu.saerok_server.domain.auth.core.entity.SocialProviderType;
 
 public interface SocialAuthClient {
